Migrated all components related to the Product detail component to it's own folder. Migrated all component test files related to the Product detail component to it's own folder. Updated references in test files and in main app component to match changes.
