+++ client/src/PDComponents/MainImageGallery.jsx
