+++ client/src/PDComponents/ProductDetailFooter.jsx
