+++ client/src/components/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import ProductDetailPage from './ProductDetailPage.jsx'
+import ProductDetailPage from '../PDComponents/ProductDetailPage.jsx'
 import FourOhFour from './FourOhFour.jsx'
 import Carousel from './Carousel.jsx'
 import ReviewSection from './Review Widget/ReviewSection.jsx'
